responsiveness adjustment using vh and px sizing format and new text files for in genereal explanation
1. Tried adjusting the responsiveness using just vh elements the last time, it was limited to only devices that are good till 574px and between 1200px to 1700px due to which tabs and some laptops were not covered in the responsiveness, however the way the website was designed earlier it was good for fevices above 992px so, set the min screen size value to 992 and put in it the css from the px sizing format and made the code working nicely for devices above 991px, however all the lower sized devices are covered by the vh sizing format. This still does not covers devices ranging between 575 and 992px. So, i will make the adjustments for those manually.

Also, commented out the card class in projects section as it was comming in the way during size adjustments.

Lastly, made new files named working.txt that has the px adjustment css. And, a tesponsiveness quotes.txt which has some messages to myself to adjust the code for diffferent sizes when i was checking the responsive behaviour of the site to different px in the inspect sectionn of chrome.